.history/backend/foodgram/api/urls_20230307140211.py
from django.urls import path
from rest_framework.authtoken import views
from users.views import registration

urlpatterns = [
    path('auth/token/login', views.obtain_auth_token),
    path('users/', registration, name='registration')
]
